fix:BoardService 삭제하고 Command와 Query 클래스로 분리
